Search for annotations

Search for annotations

search_annotations(
  limit = 20L,
  sort = c("updated", "created", "group", "id", "user"),
  search_after = NULL,
  offset = 0,
  order = c("desc", "asc"),
  uri = NULL,
  uri.parts = NULL,
  wildcard_uri = NULL,
  user = NULL,
  group = NULL,
  tags = NULL,
  any = NULL,
  quote = NULL,
  references = NULL,
  text = NULL,
  api_path = get_hs_host(),
  api_key = NULL
)

Source

https://h.readthedocs.io/en/latest/api-reference/v1/

Arguments

limit

The maximum number of annotations to return.

sort

Available values : updated, created, group, id, user, Default: `updated`

search_after

Define a start point for a subset (page) of annotation search results. Example: 2023-01-01T10:00:00.000000+00:00

offset

The number of initial annotations to skip in the result set.

order

The order in which the results should be sorted. Default: `desc`

uri

Limit the results to annotations matching the specific URI or equivalent URIs. URI can be a URL (a web page address) or a URN representing another kind of resource such as DOI (Digital Object Identifier) or a PDF fingerprint.

uri.parts

Limit the results to annotations containing the given keyword (tokenized chunk) in the URI.

wildcard_uri

Limit the results to annotations whose URIs match the wildcard pattern.

user

Limit the results to annotations made by the specified user. Example: `user=acct:username@hypothes.is`

group

Limit the results to annotations made in the specified group (by group ID).

tags

Limit the results to annotations tagged with the specified value. Example: tags=artificial,intelligence

any

Limit the results to annotations who contain the indicated keyword in any of the following field: `quote, tags, text, url`

quote

Limit the results to annotations that contain this text inside the text that was annotated.

references

Returns annotations that are replies to this parent annotation ID.

text

Limit the results to annotations that contain this text in their textual body.

api_path

The hypothesis API path, can be specify by `hypothesis.api.api_path` option or `HYPOTHESIS_API_PATH` environment variable. Default: `https://hypothes.is/api/`.

api_key

User api key, generated on the platform.